Adaptive UI: Fix circular palette reference
