+++ packages/adaptive-ui/src/color/index.ts
@@ -1,6 +1,7 @@
 export * from "./recipes/index.js";
 export * from "./utilities/index.js";
 export * from "./color.js";
+export * from "./palette-base.js";
 export * from "./palette-okhsl.js";
 export * from "./palette-rgb.js";
 export * from "./palette.js";
